Change the behavior for minisite advance search
The admin person can select available values for certain fields such
as categories and audiences. This produce a narrowed down version of
the form.
We now use this narrowed down version of the form in the advanced
search page. This is made possible be sharing in a mixin, the logic
that dynamically get and update the form fields.
