+++ src/minisites/views.py
@@ -3,9 +3,10 @@ from django.views.generic import TemplateView
 from django.contrib.sites.models import Site
 
 from search.models import SearchPage
-from aids.models import Aid
 from aids.views import SearchView, AdvancedSearchView, AidDetailView
 from alerts.views import AlertCreate
+
+from .mixins import NarrowedFiltersMixin
 
 
 class MinisiteMixin:
@@ -90,7 +91,7 @@ class MinisiteMixin:
         return context
 
 
-class SiteHome(MinisiteMixin, SearchView):
+class SiteHome(MinisiteMixin, NarrowedFiltersMixin, SearchView):
     """A static search page with admin-customizable content."""
 
     template_name = 'minisites/search_page.html'
@@ -107,55 +108,6 @@ class SiteHome(MinisiteMixin, SearchView):
         kwargs = super().get_form_kwargs()
         kwargs['data'] = data
         return kwargs
-
-    def get_available_categories(self):
-        """Return the list of categories available in this minisite.
-        Available categories are the one we select in the SearchPage admin
-        page.
-
-        When the available categories are defined in the admin, we will
-        display them in search form filter.
-
-        There is an initial filtering that applies on the list of aids when
-        loading the minisite page. That initial filtering is based on the
-        querystring field and is not affected by the selection of available
-        categories.
-        """
-        if not hasattr(self, 'available_categories'):
-            page_categories = self.search_page \
-                .available_categories \
-                .select_related('theme')
-            self.available_categories = page_categories
-        return self.available_categories
-
-    def get_available_audiences(self):
-        """Return the list of audiences available in this minisite."""
-
-        all_audiences = list(Aid.AUDIENCES)
-        available_audiences = self.search_page.available_audiences or []
-        filtered_audiences = [
-            audience for audience in all_audiences
-            if audience[0] in available_audiences
-        ]
-        return filtered_audiences
-
-    def get_form(self, form_class=None):
-        """Returns the aid search and filter form.
-
-        The minisite feature allows admin to filter the available values for
-        some filters (audiences and categories).
-        """
-        form = super().get_form(form_class)
-
-        # Show available values in categories filter field
-        available_categories = self.get_available_categories()
-        form.fields['categories'].queryset = available_categories
-
-        # Show available values in the targeted audience filter field
-        available_audiences = self.get_available_audiences()
-        form.fields['targeted_audiences'].choices = available_audiences
-
-        return form
 
     def get_queryset(self):
         """Filter the queryset on the categories and audiences filters."""
